Update ABS PKGBUILD
The Makefile that comes with ABS uses the following command for the
basic build:
```
go build -o build/abs main.go
```

This leaves behind local path information in the binary.  To remove
those references, edit the build command to pass the `-trimpath` flag to
`go build`.  This is only for the PKGBUILD.